add testcase for percentile metric aggregation

+++ elastalert/ruletypes.py
@@ -1192,9 +1192,9 @@ class MetricAggregationRule(BaseAggregationRule):
     def crossed_thresholds(self, metric_value):
         if metric_value is None:
             return False
-        if 'max_threshold' in self.rules and int(metric_value) > self.rules['max_threshold']:
+        if 'max_threshold' in self.rules and float(metric_value) > self.rules['max_threshold']:
             return True
-        if 'min_threshold' in self.rules and int(metric_value) < self.rules['min_threshold']:
+        if 'min_threshold' in self.rules and float(metric_value) < self.rules['min_threshold']:
             return True
         return False
